pyproject.toml
[tool.poetry]
name = "test-task-ketler"
version = "0.1.0"
description = ""
authors = ["MishaHmilenko <mi.khmilenko@gmail.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "^24.1.0"
aiohttp = "^3.10.8"
transformers = "^4.45.1"
pillow = "^10.4.0"
torch = "^2.4.1"
torchvision = "^0.19.1"
torchaudio = "^2.4.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
